backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
openai==1.50.0
anthropic==0.39.0
aiosqlite==0.20.0
jinja2==3.1.4
python-dotenv==1.0.1
python-multipart==0.0.9
aiofiles==24.1.0
python-docx==1.1.2
PyPDF2==3.0.1
tiktoken==0.12.0
httpx==0.27.0
